+++ copying_task.py
@@ -10,8 +10,8 @@ import time
 import os
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'  # or any {'0', '1', '2'}
 
+from EURNNOrig import EURNNOrigCell
 from EURNN import EURNNCell
-from EURNNIvan import EURNNIvanCell
 
 def copying_data(T, n_data, n_sequence):
         seq = np.random.randint(1, high=9, size=(n_data, n_sequence))
@@ -58,8 +58,8 @@ def main(model, T, n_iter, n_batch, n_hidden, capacity, comp, FFT):
                         hidden_out = tf.real(hidden_out_comp)
                 else:
                         hidden_out, _ = tf.nn.dynamic_rnn(cell, input_data, dtype=tf.float32)
-        elif model == "EURNNIvan":
-                cell = EURNNIvanCell(n_hidden, capacity, FFT, comp)
+        elif model == "EURNNOrig":
+                cell = EURNNOrigCell(n_hidden, capacity, FFT, comp)
                 if comp:
                         hidden_out_comp, _ = tf.nn.dynamic_rnn(cell, input_data, dtype=tf.complex64)
                         hidden_out = tf.real(hidden_out_comp)
