--- wholeTesting.sh
@@ -1,1 +1,0 @@
-bash testLSTM.sh 200 && bash testLSTM.sh 500 && bash testLSTM.sh 800 && bash testLSTM.sh 1000
